Added; -std=c99 option to make file. Got to replace all ushort/uint with unsigned short/unsigned int.

+++ lsf/include/lib/hdr.h
@@ -36,11 +36,11 @@
 struct LSFHeader
 {
     size_t length;
-    ushort opCode;
-    ushort version;
-    ushort reserved0;
+    unsigned short opCode;
+    unsigned short version;
+    unsigned short reserved0;
     char   padding[2];
-    uint   reserved;
+    unsigned short   reserved;
     pid_t  refCode; 
 
 };
